--- src/Captura.Native/Enums.cs
@@ -1,24 +1,0 @@
-﻿namespace Captura.Native
-{
-    public enum GetWindowEnum { Owner = 4 }
-
-    public enum SetWindowPositionFlags
-    {
-        ShowWindow = 0x400,
-        NoActivate = 0x0010
-    }
-
-    public enum WindowStyles : long
-    {
-        Child = 0x40000000,
-        ToolWindow = 0x00000080,
-        AppWindow = 0x00040000,
-        SizeBox = 0x00040000L
-    }
-
-    public enum GetWindowLongValue
-    {
-        Style = -16,
-        ExStyle = -20
-    }
-}
