+++ src/Captura.Native/Enums/GetWindowEnum.cs
@@ -1,0 +1,7 @@
+﻿namespace Captura.Native
+{
+    public enum GetWindowEnum
+    {
+        Owner = 4
+    }
+}
