+++ src/Captura.Native/Enums/GetWindowLongValue.cs
@@ -1,0 +1,8 @@
+﻿namespace Captura.Native
+{
+    public enum GetWindowLongValue
+    {
+        Style = -16,
+        ExStyle = -20
+    }
+}
