+++ src/Captura.Native/Enums/SetWindowPositionFlags.cs
@@ -1,0 +1,8 @@
+﻿namespace Captura.Native
+{
+    public enum SetWindowPositionFlags
+    {
+        ShowWindow = 0x400,
+        NoActivate = 0x0010
+    }
+}
