+++ src/Captura.Native/Enums/WindowStyles.cs
@@ -1,0 +1,10 @@
+﻿namespace Captura.Native
+{
+    public enum WindowStyles : long
+    {
+        Child = 0x40000000,
+        ToolWindow = 0x00000080,
+        AppWindow = 0x00040000,
+        SizeBox = 0x00040000L
+    }
+}
