+++ src/Captura.Native/Structs/CursorInfo.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Drawing;
+using System.Runtime.InteropServices;
+// ReSharper disable FieldCanBeMadeReadOnly.Global
+
+namespace Captura.Native
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct CursorInfo
+    {
+        public int cbSize;
+        public int flags;
+        public IntPtr hCursor;
+        public Point ptScreenPos;
+    }
+}
