+++ src/Captura.Native/Structs/RECT.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Runtime.InteropServices;
+
+namespace Captura.Native
+{
+    [Serializable, StructLayout(LayoutKind.Sequential)]
+    // ReSharper disable once InconsistentNaming
+    public struct RECT
+    {
+        public int Left;
+        public int Top;
+        public int Right;
+        public int Bottom;
+
+        public RECT(int Dimension)
+        {
+            Left = Top = Right = Bottom = Dimension;
+        }
+    }
+}
